move variant annotation to own files
