+++ lhc/binf/variant/__init__.py
@@ -1,4 +1,5 @@
 from .variant import Variant
-from .coding_variant import CodingVariant
-from .codon_variant import CodonVariant
-from .amino_acid_variant import AminoAcidVariant
+from .coding_variant import CodingVariant, call_coding_variants
+from .codon_variant import CodonVariant, call_codon_variants
+from .amino_acid_variant import AminoAcidVariant, call_amino_acid_variants
+from .variant_effect import call_variant_effects
